% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_add_images.R
\name{cm_add_images}
\alias{cm_add_images}
\title{cm_add_images}
\usage{
cm_add_images()
}
\description{
Provides an interface to examine georeferenced records.
}
\examples{
if (interactive()) {

   collmaps::cm_add_images()

}
}
