% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCollMaps.R
\name{makeCollMaps}
\alias{makeCollMaps}
\title{makeCollMaps}
\usage{
makeCollMaps()
}
\description{
Provides an interface to examine georeferenced records.
}
\examples{
if (interactive()) {

   collmaps::makeCollMaps()

}
}
